Synchronize reads and writes to svc-auth-gss global variables
This change is a part of the larger change-list for enabling
dynamic updation of NFS_KRB5 section of Ganesha configuration,
without requiring a restart of the Ganesha process. This
updation supports disabling the existing enabled rpcsec_gss
authentication, and vice-versa.

Since we intend to reset the global svc-auth-gss variables
through Ganesha, this read,write synchronization prevents misuse.

Change-Id: If3abde98799e650e90320879735d382a4f85cd78