clnt_vc: Add function to get rpc_client's service-xprt handle
When an rpc_client handle is created for a connection, the
connection's transport handle's reference is maintained within
the client duplex-receiver. This CL adds a function to return
this transport handle for a given rpc_client.

One of the applications of this function is to check if an
rpc_client used by the caller, has its transport handle being
destroyed.

BUG=b/302104182

Change-Id: If22fa39faf55b16d5f2a55c48b26e6d83f09877f