svc_vc: During connection closure, reset FD after freeing xprt's user-data
When terminating a xprt connection, the connection is first closed,
the corresponding FD is then reset, and the xprt's user-data is then
freed. This CL changes the sequence of the last two operations to
reset the connection FD only after the xprt's user-data is first
freed. This allows performing operations on xprt's user-data that
rely on the xprt FD. A sample operation could be to resolve some
connection-specific user-data that is referenced by the xprt FD.
As part of the connection termination, this data needs to be first
resolved and then released.

Change-Id: I6153eaf93e13e582e1a98210d9ac6a017b3d0981