svc_vc: Introduce flag to not create new xprt on lookup
Until now, given a FD, svc_fd_ncreatef API looks-up or
creates a xprt. The xprt creation may not be always desired.
For example, clnt_vc_ncreate_svc API operates on an existing
xprt, and it does not want svc_fd_ncreatef to create a
new xprt even if the existing xprt gets removed from the
global xprt-tree at the time of lookup.

Change-Id: I629454ca9832d861bf8ca2e9f92031569335d43e