svc_xprt/rqst: Adding xp_unique_id field to xprt
Each new xprt is assigned with a unique id (xp_unique_id).
The xprt_unique_id is set in the epoll event, and is compared with the
xprt that was found in the lookup function to verify that the current
event is indeed of the xprt that was set in the hook.

This mechanism addresses a race case scenario in which fd X is closed and
a newly created socket is assigned with the same fd X.
On the old socket an event was received but the thread was scheduled only
after a new XPRT was created with the new fd X.
The old thread handling the event might interfere with another thread
handling events for the new socket.

b/354060732

Change-Id: Ifeb43c7b4075d2f1380c647e24e85e41c9ac07c2