vc_xprt, svc_rqst: Adding xp_unique_id to xprt struct
Each new xprt is assigned with a unique id (xp_unique_id).
The xprt_unique_id is set in the epoll event, and is compared with the
xprt that was found in the lookup function to verify that the current
event is indeed of the xprt that was set in the hook.

b/354060732

Change-Id: I270cc1006055c7ae91e15f1941ee8120d5fda113