Destroy xprt when epoll register fails
When an epoll register fails we are at an undefined state.
if we just clear the connection we can leave a stuck xprt, as it is not destroyed but all its messages are cleared.
the correct solution is to destroy it and wait for a new xprt to be created same as in write failiure

b/372464776

Change-Id: If8b2c5a7d294a7090cd8fbadfb93fef338f4fc90